refactor: Удаление предупреждений

+++ src/application/logfile.cpp
@@ -40,7 +40,7 @@ void LogFile::msgHandler(QtMsgType type, const QMessageLogContext &, const QStri
   file.setFileName(mainApp->dataDir() + "/debug.log");
   QIODevice::OpenMode openMode = QIODevice::WriteOnly | QIODevice::Text;
 
-  if (file.exists() && (file.size() < maxLogFileSize)) {
+  if (file.exists() && (file.size() < (qint64)maxLogFileSize)) {
     openMode |= QIODevice::Append;
   }
